pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "longport-quant"
version = "0.1.0"
description = "Automated trading framework for Longport OpenAPI with PostgreSQL persistence"
authors = [
  {name = "Your Name", email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.25",
  "pydantic>=2.6",
  "pydantic-settings>=2.0",
  "sqlalchemy>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.12",
  "redis>=5.0",
  "tenacity>=8.2",
  "loguru>=0.7",
  "apscheduler>=3.10",
  "python-dotenv>=1.0",
  "PyYAML>=6.0",
  "longport>=3.0.14",
  "numpy>=1.24",
  "pandas>=2.0",
  "ta>=0.10"
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
  "pytest-asyncio>=0.23",
  "ruff>=0.4",
  "mypy>=1.7"
]

[project.urls]
Homepage = "https://github.com/your-org/longport-quant"
Documentation = "https://open.longportapp.com/zh-CN/docs"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
select = ["E", "F", "B", "C4"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra"
